quarantine: keep quarantine off by default
allow path specification for the quarantine directory

+++ wsd/QuarantineUtil.cpp
@@ -21,6 +21,11 @@ namespace Quarantine
 {
     void createQuarantineMap()
     {
+        bool enable = LOOLWSD::getConfigValue<bool>("quarantine_files[@enable]", false);
+
+        if (!enable)
+            return;
+
         std::vector<std::string> files;
         Poco::File(LOOLWSD::QuarantinePath).list(files);
         LOOLWSD::QuarantineMap.clear();
@@ -43,6 +48,11 @@ namespace Quarantine
 
     void removeQuarantine()
     {
+        bool enable = LOOLWSD::getConfigValue<bool>("quarantine_files[@enable]", false);
+
+        if (!enable)
+            return;
+
         FileUtil::removeFile(LOOLWSD::QuarantinePath, true);
     }
 
@@ -51,6 +61,10 @@ namespace Quarantine
     // because they are originally stored in jails
     std::size_t quarantineSize()
     {
+        bool enable = LOOLWSD::getConfigValue<bool>("quarantine_files[@enable]", false);
+
+        if (!enable)
+            return 0;
         std::vector<std::string> files;
         Poco::File(LOOLWSD::QuarantinePath).list(files);
         std::size_t size = 0;
@@ -66,10 +80,12 @@ namespace Quarantine
 
     void makeQuarantineSpace()
     {
-        std::size_t sizeLimit = LOOLWSD::getConfigValue<std::size_t>("quarantine_files.limit_dir_size_mb", 0)*1024*1024;
+        bool enable = LOOLWSD::getConfigValue<bool>("quarantine_files[@enable]", false);
 
-        if(sizeLimit == 0)
+        if (!enable)
             return;
+
+        std::size_t sizeLimit = LOOLWSD::getConfigValue<std::size_t>("quarantine_files.limit_dir_size_mb", 0)*1024*1024;
 
         std::vector<std::string> files;
         Poco::File(LOOLWSD::QuarantinePath).list(files);
@@ -94,6 +110,11 @@ namespace Quarantine
 
     void clearOldQuarantineVersions(std::string Wopiscr)
     {
+        bool enable = LOOLWSD::getConfigValue<bool>("quarantine_files[@enable]", false);
+
+        if (!enable)
+            return;
+
         std::size_t maxVersionCount = LOOLWSD::getConfigValue<std::size_t>("quarantine_files.max_versions_to_maintain", 2);
         std::string decoded;
         Poco::URI::decode(Wopiscr, decoded);
@@ -107,10 +128,9 @@ namespace Quarantine
 
     bool quarantineFile(DocumentBroker* docBroker, std::string docName)
     {
-        //return if directory size limit is 0
-        std::size_t sizeLimit = LOOLWSD::getConfigValue<std::size_t>("quarantine_files.limit_dir_size_mb", 0);
+        bool enable = LOOLWSD::getConfigValue<bool>("quarantine_files[@enable]", false);
 
-        if (sizeLimit == 0)
+        if (!enable)
             return false;
 
         std::string docKey;
